% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMaNam.R
\name{calc_occurrence}
\alias{calc_occurrence}
\title{Create a data format for marker occurrences, which will be used to identify and remove the most common markers across the dataset.}
\usage{
calc_occurrence(markers_data, features_col = NULL, cell_column = NULL)
}
\arguments{
\item{markers_data}{Dataframe containing marker data}

\item{features_col}{Name of the column containing features}

\item{cell_column}{Name of the column containing cell annotations}
}
\value{
Dataframe with feature occurrence counts
}
\description{
Create a data format for marker occurrences, which will be used to identify and remove the most common markers across the dataset.
}
