% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMaNam.R
\name{cell_typing}
\alias{cell_typing}
\title{Perform cell type annotation based on annotation (get_annotation()).data.
If necessary, you can use hierarchical data (via `load_markers_hierarchy()`) to find and construct hierarchical relationships in cell names.}
\usage{
cell_typing(
  annotation_data,
  hierarchy_data,
  p_val = 0.01,
  level = 1,
  hierarchy = TRUE
)
}
\arguments{
\item{annotation_data}{Dataframe with cell annotations and p-values (get_annotation())}

\item{hierarchy_data}{Dataframe representing the hierarchy of cell types (load_markers_hierarchy() or NULL if not needed)}

\item{p_val}{Threshold p-value for selecting significant annotations. Default: 0.01}

\item{level}{Integer, specifying the level in the hierarchy to use. Default: 1}

\item{hierarchy}{Boolean, whether to use hierarchical annotation. If you want to use hierarchical data for name creating. If not the names will based on 'level'. Default: TRUE}
}
\value{
A dataframe with assigned cell types, weighted p-values, and log-transformed p-values
}
\description{
Perform cell type annotation based on annotation (get_annotation()).data.
If necessary, you can use hierarchical data (via `load_markers_hierarchy()`) to find and construct hierarchical relationships in cell names.
}
\details{
Uses hierarchical relationships and statistical thresholds to classify cell types.
}
