% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMaNam.R
\name{find_markers_rna_seq}
\alias{find_markers_rna_seq}
\title{Identify markers from RNA sequencing (scRNAseq) data.
The input dataframe should be tailored to the type of RNA sequencing used:
For single-cell RNA sequencing (scRNA-seq) where each cell is sorted and sequenced like RNAseq each column should represent an individual cell.
In the case of Drop-seq data with cluster information, the dataframe should include the average expression per cluster, with each cluster represented by its name.}
\usage{
find_markers_rna_seq(
  rna_seq_df,
  exclude_genes = NULL,
  exclude_mt = TRUE,
  include_markers = NULL,
  stat = "mean"
)
}
\arguments{
\item{rna_seq_df}{Dataframe with RNA sequencing (scRNAseq) data}

\item{exclude_genes}{List of genes to exclude}

\item{exclude_mt}{Boolean, whether to exclude mitochondrial genes. Default: TRUE}

\item{include_markers}{List of markers to include. Recommended all markers inside in cell markers data used for naming.}

\item{stat}{Statistical method for filtering ('mean', 'median', 'q1', 'q3'). Default: 'mean'}
}
\value{
Dataframe of identified markers
}
\description{
Identify markers from RNA sequencing (scRNAseq) data.
The input dataframe should be tailored to the type of RNA sequencing used:
For single-cell RNA sequencing (scRNA-seq) where each cell is sorted and sequenced like RNAseq each column should represent an individual cell.
In the case of Drop-seq data with cluster information, the dataframe should include the average expression per cluster, with each cluster represented by its name.
}
