% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMaNam.R
\name{get_annotation}
\alias{get_annotation}
\title{Annotate cell markers based on statistical significance.
This function annotates each cell using the provided markers for cell naming, based on the markers returned by `find_markers_rna_seq()`.
Alternatively, you can use a different set of markers (e.g., obtained from Seurat) by loading them into the function.
For custom cell marker sets, you must modify the 'cell_markers' dataframe, ensuring that the cell names are in the 'cell_annotation' column and the genes are listed under a separate 'features' column.}
\usage{
get_annotation(cell_markers, markers_occ, max_genes)
}
\arguments{
\item{cell_markers}{Dataframe containing marker information for each cell}

\item{markers_occ}{Dataframe with marker occurrences in different cell types}

\item{max_genes}{Integer, the maximum number of genes to consider}
}
\value{
A dataframe with cell annotations, associated cell types, p-values, and gene percentages
}
\description{
Annotate cell markers based on statistical significance.
This function annotates each cell using the provided markers for cell naming, based on the markers returned by `find_markers_rna_seq()`.
Alternatively, you can use a different set of markers (e.g., obtained from Seurat) by loading them into the function.
For custom cell marker sets, you must modify the 'cell_markers' dataframe, ensuring that the cell names are in the 'cell_annotation' column and the genes are listed under a separate 'features' column.
}
\details{
Uses Fisher's exact test to assess marker significance.
}
