% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMaNam.R
\name{heat_map_names}
\alias{heat_map_names}
\title{Generate a heatmap of cell type annotations}
\usage{
heat_map_names(cell_types_data, value = "completed")
}
\arguments{
\item{cell_types_data}{Dataframe containing cell type data (cell_typing())}

\item{value}{Column name used for heatmap values. Default: 'completed'}
}
\value{
A heatmap object displaying the relationship between full names and annotations
}
\description{
Generate a heatmap of cell type annotations
}
