% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMaNam.R
\name{load_markers_data}
\alias{load_markers_data}
\title{Load marker data from a file
Marker data structure available on GitHub (https://github.com/jkubis96/CellMaNam)}
\usage{
load_markers_data(data_path)
}
\arguments{
\item{data_path}{Path to the input file (CSV, TSV, TXT, or XLSX)}
}
\value{
A dataframe containing the marker data
}
\description{
Load marker data from a file
Marker data structure available on GitHub (https://github.com/jkubis96/CellMaNam)
}
