% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMaNam.R
\name{select_top_occ}
\alias{select_top_occ}
\title{Identify the most frequently occurring features for each cell annotation.
The input data must be derived from the calc_occurrence() function.}
\usage{
select_top_occ(markers_occ, top_n = 2)
}
\arguments{
\item{markers_occ}{Dataframe containing feature occurrences}

\item{top_n}{Number of thresholds for selecting the top features}
}
\value{
Filtered dataframe with top occurring features
}
\description{
Identify the most frequently occurring features for each cell annotation.
The input data must be derived from the calc_occurrence() function.
}
